fix: add missing n_channels in denominator term for 3D and 4D multichannel
The 2D case already had n_channels in the denominator of the argument to
fast_exp, but this was missing for 3D and 4D.  It should be there in all cases
to get consistent denoising regardless of the number of channels.  (i.e. the
user would not need to tune `h` depending on the number of channels).